Update README.md
I replaced the string data types with bytes32, which is a fixed-size byte array. We've also reduced the size of the uint data types to uint16 where appropriate, since we know that these values will never exceed the maximum value of uint16.

These changes can help to reduce the overall gas cost of the contract, making it more efficient to use. However, it's important to ensure that these changes don't compromise the functionality or security of the contract.